Get rid of ability to register a user
Add an admin dashboard that requires authentication to get in.

We are getting rid of registration because there will only ever be one admin on the site (me).
Removing registration also allowed me to remove confirmation because those two flows are linked.
And we don't need either.

Finally, we have an admin dashboard where you will be writing posts.